+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 0.1.1 (unreleased)
 ------------------
 
+* If unset, `resources` now provide an AWS account as the default resource 
 * All datetime attributes are properly validated for ISO8601 format
+
 
 0.1.0 (2020-12-13)
 -------------------